Blade-202: Use NIO, remove usage of aQute.IO where possible

+++ src/main/java/com/liferay/blade/cli/util/FilesUtil.java
@@ -1,20 +1,54 @@
 package com.liferay.blade.cli.util;
 
 import java.io.IOException;
-import java.nio.file.FileVisitOption;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.Comparator;
 import java.util.stream.Collectors;
 
+import com.liferay.blade.cli.CopyDirVisitor;
+
+import java.nio.file.*;
+
 public class FilesUtil {
 
-	public static void deleteWithException(Path path) throws IOException {
-		for (Path p : Files.walk(path, FileVisitOption.FOLLOW_LINKS)
-		.sorted(Comparator.reverseOrder())
-		.collect(Collectors.toList())) {
-			Files.delete(p);
+	public static void delete(Path path) throws IOException {
+		if (Files.exists(path))
+		{
+			if (Files.isDirectory(path))
+			{
+				for (Path p : Files.walk(path, FileVisitOption.FOLLOW_LINKS).sorted(Comparator.reverseOrder())
+						.collect(Collectors.toList())) {
+					Files.delete(p);
+				}
+				
+			}
+			else
+			{
+				Files.delete(path);
+			}
 		}
+	}
+
+	public static void copy(Path sourceDir, Path targetDir) throws IOException {
+		Files.walkFileTree(sourceDir, new CopyDirVisitor(sourceDir, targetDir, StandardCopyOption.REPLACE_EXISTING));
+	}
+	
+	public static String convertStreamToString(java.io.InputStream is) {
+	   
+		final String returnValue;
+		try (java.util.Scanner s = new java.util.Scanner(is))
+	    {
+			s.useDelimiter("\\A");
+			returnValue = s.hasNext() ? s.next() : "";
+	    }
+	    finally {
+	    	try {
+	    		is.close();
+	    	}
+	    	catch (Exception e) {
+	    		e.printStackTrace();
+	    	}
+	    }
+		return returnValue;
 	}
 
 }
