+++ src/test/java/com/liferay/blade/cli/gradle/GradleToolingTest.java
@@ -24,12 +24,12 @@ import static org.junit.Assert.assertTrue;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
 import java.util.Set;
 
 import org.junit.BeforeClass;
 import org.junit.Test;
 
+import com.liferay.blade.cli.util.FilesUtil;
 import com.liferay.project.templates.internal.util.FileUtil;
 
 /**
@@ -39,12 +39,13 @@ public class GradleToolingTest {
 
 	@BeforeClass
 	public static void setUpClass() throws Exception {
-		Files.copy(Paths.get("deps.zip"), Paths.get("build","classes","java","test","deps.zip"), StandardCopyOption.REPLACE_EXISTING);
+		FilesUtil.copy(Paths.get("deps.zip"), Paths.get("build","classes","java","test","deps.zip"));
 		Path testws = Paths.get("build","testws1");
-		FileUtil.deleteDir(testws);
+		if (Files.exists(testws))
+			FileUtil.deleteDir(testws);
 		Files.createDirectories(testws);
-		Files.copy(
-			Paths.get("test-resources","projects","testws1"), testws, StandardCopyOption.REPLACE_EXISTING);
+		FilesUtil.copy(
+			Paths.get("test-resources","projects","testws1"), testws);
 	}
 
 	@Test
